refactor(module_utils): give PgMembership a module_utils of its own
The class was 439 of the 921 lines of a file every module in the collection
imports, with exactly one consumer. Move it to module_utils/membership.py, which
leaves postgres.py to the connection and query helpers it shares.

Derive the name of each membership option from one constant. The keyword
PostgreSQL uses in GRANT, the pg_auth_members column and the module parameter
differ only by case and a suffix, but were restated separately at each call site
and had already drifted in spelling. They cannot drift now.
