+++ changelogs/fragments/postgresql_membership-granting-role.yml
@@ -5,6 +5,7 @@ breaking_changes:
 minor_changes:
   - postgresql_membership - add the ``admin_option``, ``inherit_option`` and ``set_option`` options, controlling the membership options of the grants the module makes on PostgreSQL 16 and later (https://github.com/ansible-collections/community.postgresql/issues/757).
   - postgresql_membership - add the ``granted_by`` option, naming the role to record as the granting role. It is required when the connecting role holds ``ADMIN OPTION`` only indirectly, and manages a grant made by a different role.
+  - postgresql_membership - move ``PgMembership`` out of the shared ``postgres`` module_utils into a ``membership`` module_utils of its own, and derive the name of every membership option from one constant instead of restating it per call site. No behaviour change; an external importer of ``PgMembership`` has to follow the move.
   - postgresql_membership - add the ``grants`` and ``effective_options`` return values, reporting every grant of the requested pairs and the options the target role effectively holds.
   - postgresql_membership - ``state=exact`` now reports ``granted`` with an entry for every requested group, as ``state=present`` already did. It previously listed only the groups that changed, so a task asserting ``result.granted == {}`` on an unchanged run now sees one empty entry per requested group.
